% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{do}
\alias{do}
\title{Call a function in shorthand.}
\usage{
do(x = NULL, f)
}
\arguments{
\item{x}{(Default: \code{NULL}) An object passed into \code{f} if \code{f} is not parameterless.}

\item{f}{A one-argument or parameterless function. If one-argument, then \code{f(x)} is returned. Otherwise, \code{f()} is returned.}
}
\value{
The output of \code{f(x)} or \code{f()}, whatever that may be.
}
\description{
A convenience function that either: a) applies a function to an input or b) runs a parameterless function. A useful shorthand alternative to piping into an anonymous function (a) or manually removing intermediate objects used in creating an output (b).
}
\examples{
library("tidyverse")
library("PipeHelpR")

# make some test data
set.seed(123)
testvector = rnorm(100, mean=10, sd=3)

# standardize the vector to mean 0 variance 1
testvector \%>\%
  # substract off the arithmetic mean
  {\(x)x - mean(x)}() \%>\%
  # divide by empirical standard deviation
  {\(y)y / sd(y)}() \%>\%
  # limit output
  head()
# > [1] -0.71304802 -0.35120270  1.60854170 -0.02179795  0.04259548  1.77983218

# do the same but using do() for better readability
testvector \%>\%
  do(\(x)x - mean(x)) \%>\%
  do(\(y)y / sd(y)) \%>\%
  head()
# > [1] -0.71304802 -0.35120270  1.60854170 -0.02179795  0.04259548  1.77983218
}
