% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xindex.R
\name{index}
\alias{index}
\title{Perform general indexing in a pipeline.}
\usage{
index(object, idx = NA, neg = F)
}
\arguments{
\item{object}{The object to perform the indexing upon.}

\item{idx}{If \code{object} is unidimensional, an integer/character/logical vector--or singleton list with such a vector--to use in indexing \code{object}; otherwise, a list of integer/character/logical vectors to use in indexing each dimension. To keep all indices in a certain dimension, use \code{NA}.}

\item{neg}{(Default: \code{FALSE}) A logical vector that determines whether to use negative indexing or not. If \code{object} is multidimensional, negative indexing can be used for certain dimensions and not others by making this argument a vector.}
}
\value{
A copy of \code{object} indexed accordingly.
}
\description{
A convenience function that performs indexing on many general objects with many possible methods, and allows for negative indexing regardless of indexing type.
}
\examples{
library("tidyverse")
library("PipeHelpR")

# make some test data
set.seed(123)
testmatrix = rnorm(100, mean=10, sd=3) \%>\%
  # make into 10x10 matrix
  matrix(nrow=10) \%>\%
  setRownames(LETTERS[1:10]) \%>\%
  setColnames(LETTERS[11:20])

# indexing can be done in a pipeline
testmatrix \%>\%
  `+`(10) \%>\%
  index(list(c("A","C"), NA)) # NA is used for an empty index ("get all columns")
# >          K        L        M        N        O        P        Q        R        S        T
# > A 18.31857 23.67225 16.79653 21.27939 17.91588 20.75996 21.13892 18.52691 20.01729 22.98051
# > C 24.67612 21.20231 16.92199 22.68538 16.20381 19.87139 19.00038 23.01722 18.88802 20.71620

# negative indexing can be done with characters!
testmatrix \%>\%
  `*`(2) \%>\%
  # negative indexing for the first dimension, normal indexing for the second
  index(list(c("E","F","G","H","I","J"), 10:9), neg=c(T,F))
# >          T        S
# > A 25.96102 20.03459
# > B 23.29038 22.31168
# > C 21.43239 17.77604
# > D 16.23256 23.86626
}
