% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xindex.R
\name{lindex}
\alias{lindex}
\title{Perform (multi-level) list indexing in a pipeline.}
\usage{
lindex(object, idx)
}
\arguments{
\item{object}{The list to perform the indexing upon.}

\item{idx}{A list or vector of integer/character objects to use in indexing \code{object}.}
}
\value{
A copy of \code{object} indexed accordingly.
}
\description{
A convenience function that performs list single-element indexing, and permits indexing at multiple levels.
}
\examples{
library("tidyverse")
library("PipeHelpR")
testlist = list(
  a = list(c = 1, d = 2),
  b = list(e = 3, f = 4)
)
testlist \%>\% lindex("a")
# > $c
# > [1] 1
# > $d
# > [1] 2
testlist \%>\% lindex(list(2, "e"))
# > [1] 3
}
