% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setXNames.R
\name{setColnames}
\alias{setColnames}
\title{Set object \code{col.names} in a pipeline.}
\usage{
setColnames(object, nm)
}
\arguments{
\item{object}{The object whose column names will be modified.}

\item{nm}{A character vector containing the new column names for \code{object}.}
}
\value{
A copy of \code{object} with its column names changed to \code{nm}.
}
\description{
A convenience function to modify the column names of the passed in object.
}
\examples{
library("tidyverse")
library("PipeHelpR")

# make some test data
set.seed(123)
testvector = rnorm(100, mean=10, sd=3)

# make a matrix out of the vector, and add row and column names
testmatrix = testvector \%>\%
  # make into 10x10 matrix
  matrix(nrow=10) \%>\%
  setColnames(LETTERS[11:20])

head(testmatrix, c(6,6))
# >              K         L        M         N         O         P
# > [1,]  8.318573 13.672245 6.796529 11.279393  7.915879 10.759956
# > [2,]  9.309468 11.079441 9.346075  9.114786  9.376248  9.914360
# > [3,] 14.676125 11.202314 6.921987 12.685377  6.203811  9.871389
# > [4,] 10.211525 10.332048 7.813326 12.634400 16.506868 14.105807
# > [5,] 10.387863  8.332477 8.124882 12.464743 13.623886  9.322687
# > [6,] 15.145195 15.360739 4.939920 12.065921  6.630674 14.549412
}
