% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripAttr.R
\name{stripAttr}
\alias{stripAttr}
\title{Remove object attributes.}
\usage{
stripAttr(x, which)
}
\arguments{
\item{x}{An object.}

\item{which}{A character object with the attribute name(s) to remove.}
}
\value{
A copy of \code{x} with the selected attributes removed.
}
\description{
A convenience function that strips selected attributes of an object.
}
\examples{
library("tidyverse")
library("PipeHelpR")

# make some test data
set.seed(123)
testarray = rnorm(100, mean=10, sd=3) \%>\% array(dim=c(4,5,5))

# recover the underlying vector from testarray
testarray \%>\% stripAttr("dim") \%>\% head()
# > [1]  8.318573  9.309468 14.676125 10.211525 10.387863 15.145195
}
