% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichInv.R
\name{whichInv}
\alias{whichInv}
\title{Perform the functional inverse of \code{base::which()}.}
\usage{
whichInv(idx, len)
}
\arguments{
\item{idx}{An integer vector containing the \code{TRUE} indices in the output.}

\item{len}{An integer that determines the length of the output.}
}
\value{
A logical vector of length \code{len} with all \code{idx} indices set to \code{TRUE} and the rest to \code{FALSE.}
}
\description{
A convenience function that performs the reverse operation of \code{base::which()}.
}
\examples{
library("tidyverse")
library("PipeHelpR")

seq(1, 10, 2) \%>\%
  whichInv(10)
# > [1]  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE
}
