% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setXNames.R
\name{setDimnames}
\alias{setDimnames}
\title{Set object \code{dimnames} in a pipeline.}
\usage{
setDimnames(object, ...)
}
\arguments{
\item{object}{The object whose dimension names will be modified.}

\item{...}{Either a single list of character vectors or multiple separate character vectors containing the new dimension names for \code{object}.}
}
\value{
A copy of \code{object} with its dimension changed to the elements of \code{...}.
}
\description{
A convenience function to modify the dimension names of the passed in object.
}
\examples{
library("tidyverse")
library("PipeHelpR")

# make some test data
set.seed(123)
testvector = rnorm(100, mean=10, sd=3)

# make an array out of the vector, and add names for its dimensions
testarray = testvector \%>\%
  array(dim=c(4,5,5)) \%>\%
  setDimnames(LETTERS[1:4], LETTERS[5:9], LETTERS[10:14])

head(testarray, c(2,2,2))
# > , , J
# >          E        F
# > A 8.318573 10.38786
# > B 9.309468 15.14519
# >
# > , , K
# >          E        F
# > A 6.796529 8.124882
# > B 9.346075 4.939920
}
